Normalise floating point values in snapshots
The snapshot tests were failing on my machine due to minor differences in 
long floating-point values, presumably since I do not use the same platform
as these snapshots were created for.

To remedy this we unconditionally truncate any stream of 10+ numbers following
a period when generated a snapshot. XML is untyped (without us parsing the SVG
values) so this is not 100% guaranteed to find only long floating-point numbers
but it seems fairly safe.